// src/types/REST/api/generated/models/OtherSeoComponent.ts
export interface OtherSeoComponent {
    /**
     * 
     * @type {number}
     * @memberof OtherSeoComponent
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof OtherSeoComponent
     */
    metaTitle?: string;
    /**
     * 
     * @type {string}
     * @memberof OtherSeoComponent
     */
    metaDescription?: string;
    /**
     * 
     * @type {string}
     * @memberof OtherSeoComponent
     */
    keywords?: string;
    /**
     * 
     * @type {string}
     * @memberof OtherSeoComponent
     */
    metaRobots?: string;
    /**
     * 
     * @type {string}
     * @memberof OtherSeoComponent
     */
    canonicalURL?: string;
    /**
     * 
     * @type {Array<OtherSeoComponentMetaSocialsInner>}
     * @memberof OtherSeoComponent
     */
    metaSocials?: Array<OtherSeoComponentMetaSocialsInner>;
    /**
     * 
     * @type {Array<OtherSeoComponentOtherMetasInner>}
     * @memberof OtherSeoComponent
     */
    otherMetas?: Array<OtherSeoComponentOtherMetasInner>;
}